Use an iterator increment that has an exact binary number representation
